feat(syntax-highlighter): clean up components
