+++ packages/waystone-components/src/code-block/codeBlock.styles.ts
@@ -3,7 +3,6 @@ import { ThemeUIStyleObject } from 'theme-ui';
 
 export const codeBlockContainerCss: ThemeUIStyleObject = {
   backgroundColor: 'b-600',
-  px: 'md',
   mb: 'lg',
   mx: 'md',
   borderRadius: 'md',
