+++ packages/waystone-syntax-highlighter/src/components/index.ts
@@ -1,0 +1,1 @@
+export * from './code-line/codeLine';
