+++ packages/waystone-syntax-highlighter/src/hooks.ts
@@ -1,0 +1,13 @@
+import { useMemo } from 'react';
+import { CreateCodeTreeOptions, createCodeTree } from './utils/createCodeTree';
+import { flattenCodeTree } from './utils/flattenCodeTree';
+
+export const useCodeTree = (options: CreateCodeTreeOptions) => {
+  const codeTree = useMemo(() => {
+    const { children: tree } = createCodeTree(options);
+
+    return flattenCodeTree({ tree });
+  }, [options]);
+
+  return codeTree;
+};
