+++ packages/waystone-syntax-highlighter/src/index.ts
@@ -1,2 +1,3 @@
-export * from './SyntaxHighlighter';
+export * from './component';
 export * from './lib/supportedLanguages.generated';
+export * from './lib/types';
