+++ packages/waystone-syntax-highlighter/src/lib/index.ts
@@ -1,3 +1,5 @@
 export * from './highlightStyles.generated';
 export * from './supportedLanguages.generated';
-export * from './styles';
+export * from './types';
+
+export * as styles from './styles';
