+++ packages/waystone-syntax-highlighter/src/utils/createStyleObject.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from '../lib';
+import { getClassNameCombinations } from './getClassNameCombinations';
+
+export type CreateStyleObjectOptions = {
+  classNames: string[];
+  stylesheet: StyleSheet;
+};
+
+export const createStyleObject = ({
+  classNames,
+  stylesheet,
+}: CreateStyleObjectOptions) => {
+  const nonTokenClassNames = classNames.filter(
+    (className) => className !== 'token'
+  );
+  const classNamesCombinations = getClassNameCombinations(nonTokenClassNames);
+  return classNamesCombinations.reduce(
+    (styleObject, className) => ({
+      ...styleObject,
+      ...stylesheet[className as keyof typeof stylesheet],
+    }),
+    {}
+  );
+};
