+++ packages/waystone-syntax-highlighter/src/utils/index.ts
@@ -1,0 +1,8 @@
+export * from './createSpan';
+export * from './createPowerSetPermutations';
+export * from './createCodeTree';
+export * from './createStyleObject';
+export * from './flattenCodeTree';
+export * from './getClassNameCombinations';
+
+export * from '../lib/types';
